whisper_standalone_project/requirements.txt
# ==============================
# Core ML
# ==============================
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.1.99

# ==============================
# Whisper STT
# ==============================
openai-whisper>=20231117

# ==============================
# Audio processing
# ==============================
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6

# ==============================
# Translation + Romanization
# ==============================
indic-transliteration>=2.3.0

# ==============================
# ElevenLabs TTS
# ==============================
elevenlabs

# ==============================
# Utilities
# ==============================
tqdm>=4.65.0

# Transformers dependencies (important)
huggingface_hub>=0.23.0
tokenizers>=0.15.0
safetensors>=0.4.0
packaging>=23.0
pyyaml>=6.0
regex>=2023.0